prog.py
from __future__ import print_function
from sys import exit
import argparse
import os
import numpy as np
import matplotlib as mpl
import matplotlib.pyplot as plt
import matplotlib.backends.backend_agg as agg
import time
from datetime import datetime
import cv2
import pygame
import serial
import serial.tools.list_ports
import threading
import csv
import urllib.error
import urllib.request
import struct
from OpenGL.GL import *
from OpenGL.GLU import *
from OpenGL.GL.shaders import compileProgram, compileShader
import pyrr
import pdb
import gc
import functools

import hashlib
import hmac
import base64
import urllib.parse as urlparse

from PIL import Image

from tkinter import Tk
from tkinter.filedialog import askopenfilename
import groundstation_v2
